Use migration package for repo migration
