+++ cmd/migrateRepoCmd.go
@@ -1,6 +1,3 @@
-/*
-Package cmd provides a command-line interface for changing GHAS settings for a given organization.
-*/
 package cmd
 
 import (
@@ -17,22 +14,12 @@ const (
 	repositoryFlagName = "repo"
 )
 
-// migrateRepoCmd represents the migrateRepo command
 var migrateRepoCmd = &cobra.Command{
 	Use:   "migrate-repository",
 	Short: "Migrate a repository",
 	Long: `This script migrates a repositories from one organization to another.
 
-	The target organization has to exist at destination.
-
-	Migration steps:
-
-	- 1. Deactivate GHAS settings at target organization
-	- 2. Deactivate GHAS settings at source repository
-	- 3. Migrate repository
-	- 4. Delete branch protections at target
-	- 5. change repository visibility to internal at target
-	- 6. Activate GHAS settings at target`,
+	The target organization has to exist at destination.`,
 	Run: func(cmd *cobra.Command, args []string) {
 		sourceOrg, _ := cmd.Flags().GetString(sourceOrgFlagName)
 		targetOrg, _ := cmd.Flags().GetString(targetOrgFlagName)
@@ -50,7 +37,8 @@ var migrateRepoCmd = &cobra.Command{
 			os.Exit(1)
 		}
 
-		err = migration.ProcessRepoMigration(slog.Default(), repo, sourceOrg, targetOrg, sourceToken, targetToken, maxRetries)
+		migration := migration.NewRepoMigration(*repo.Name, sourceOrg, targetOrg, sourceToken, targetToken, maxRetries)
+		err = migration.Migrate()
 
 		if err != nil {
 			slog.Error("error migrating repository: " + repository)
