+++ cmd/rootCmd.go
@@ -1,6 +1,3 @@
-/*
-Copyright © 2023 NAME HERE <EMAIL ADDRESS>
-*/
 package cmd
 
 import (
@@ -33,18 +30,12 @@ func initLogger(cmd *cobra.Command, args []string) {
 	logging.NewLoggerFromContext(ctx, enableDebug)
 }
 
-// rootCmd represents the base command when called without any subcommands
 var rootCmd = &cobra.Command{
 	Use:              "gei-migration-helper",
 	PersistentPreRun: initLogger,
 	Short:            "Wrapper application to the GEI extension that orchestrates steps necessary to migrate reposistories and GHAS features",
-	// Uncomment the following line if your bare application
-	// has an action associated with it:
-	// Run: func(cmd *cobra.Command, args []string) { },
 }
 
-// Execute adds all child commands to the root command and sets flags appropriately.
-// This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute() {
 	err := rootCmd.Execute()
 
@@ -59,9 +50,6 @@ func init() {
 	rootCmd.Version = VERSION
 	rootCmd.Println("\n\n" + string(banner) + "\n\n")
 
-	// Here you will define your flags and configuration settings.
-	// Cobra supports persistent flags, which, if defined here,
-	// will be global for your application.
 	rootCmd.PersistentFlags().BoolVar(&enableDebug, "debug", os.Getenv("DEBUG") == "true", "Enable debug mode")
 
 	rootCmd.PersistentFlags().String(sourceOrgFlagName, "", "The source organization.")
