+++ internal/migration/orgMigration.go
@@ -14,21 +14,20 @@ import (
 )
 
 type OrgMigration struct {
-	source, target           string
-	sourceToken, targetToken string
-	maxRetries               int
-	parallelMigrations       int
+	parallelMigrations int
+	orgs               orgs
 }
 
 const statusRepoName = "migration-status"
 
-func NewOrgMigration(source, target, sourceToken, targetToken string, maxRetries int, parallelMigrations int) OrgMigration {
-	return OrgMigration{source, target, sourceToken, targetToken, maxRetries, parallelMigrations}
+func NewOrgMigration(source, target, sourceToken, targetToken string, retries int, parallelMigrations int) OrgMigration {
+	maxRetries = retries
+	return OrgMigration{parallelMigrations, orgs{source, target, sourceToken, targetToken}}
 }
 
 func (om OrgMigration) checkOngoing() error {
 	slog.Info("looking for ongoing/past migration")
-	repo, err := github.GetRepository(statusRepoName, om.target, om.targetToken)
+	repo, err := github.GetRepository(statusRepoName, om.orgs.target, om.orgs.targetToken)
 
 	if err != nil && err.Error() != github.ErrRepositoryNotFound.Error() {
 		slog.Error("error fetching migration status repository", err)
@@ -36,20 +35,20 @@ func (om OrgMigration) checkOngoing() error {
 	}
 
 	if err == nil && *repo.Name == statusRepoName {
-		issue, _ := github.GetIssue(om.target, statusRepoName, 1, om.targetToken)
+		issue, _ := github.GetIssue(om.orgs.target, statusRepoName, 1, om.orgs.targetToken)
 
 		if issue != nil {
-			err := fmt.Errorf("a migration to this organization was already executed. Please check http://github.com/%s/%s/issues/1 for status", om.target, statusRepoName)
+			err := fmt.Errorf("a migration to this organization was already executed. Please check http://github.com/%s/%s/issues/1 for status", om.orgs.target, statusRepoName)
 			slog.Error(err.Error())
 			return err
 		}
 
-		err := fmt.Errorf("a migration to this organization is either ongoing or finished in error (remove http://github.com/%s/%s if you want to retry)", om.target, statusRepoName)
+		err := fmt.Errorf("a migration to this organization is either ongoing or finished in error (remove http://github.com/%s/%s if you want to retry)", om.orgs.target, statusRepoName)
 		return err
 	}
 
 	slog.Info("creating migration status repository")
-	github.CreateRepository(om.target, statusRepoName, om.targetToken)
+	github.CreateRepository(om.orgs.target, statusRepoName, om.orgs.targetToken)
 
 	return nil
 }
@@ -60,7 +59,7 @@ func (om OrgMigration) prepareMigration() error {
 	}
 
 	slog.Info("deactivating GHAS settings at target organization")
-	github.ChangeGHASOrgSettings(om.target, false, om.targetToken)
+	github.ChangeGHASOrgSettings(om.orgs.target, false, om.orgs.targetToken)
 
 	return nil
 }
@@ -85,7 +84,7 @@ func (om OrgMigration) Process(repo interface{}, ctx context.Context) error {
 
 	slog.Info("starting migration", "name", *repository.Name)
 	logger := logging.NewLoggerFromContext(ctx, false)
-	err := ProcessRepoMigration(logger, repository, om.source, om.target, om.sourceToken, om.targetToken, om.maxRetries)
+	err := processRepoMigration(logger, repository, om.orgs.source, om.orgs.target, om.orgs.sourceToken, om.orgs.targetToken)
 	slog.Info("finished migrating", "name", *repository.Name)
 	if err != nil {
 		slog.Error("error migrating repository: ", err)
@@ -95,24 +94,24 @@ func (om OrgMigration) Process(repo interface{}, ctx context.Context) error {
 	return nil
 }
 
-func (om OrgMigration) Migrate() (MigrationResult, error) {
+func (om OrgMigration) Migrate() (migrationResult, error) {
 	if err := om.prepareMigration(); err != nil {
-		return MigrationResult{}, err
+		return migrationResult{}, err
 	}
 
 	slog.Info("fetching repositories from source organization")
-	sourceRepositories, err := github.GetRepositories(om.source, om.sourceToken)
+	sourceRepositories, err := github.GetRepositories(om.orgs.source, om.orgs.sourceToken)
 
 	if err != nil {
 		slog.Error("error fetching repositories from source organization")
-		return MigrationResult{}, err
+		return migrationResult{}, err
 	}
 
-	destinationRepositories, err := github.GetRepositories(om.target, om.targetToken)
+	destinationRepositories, err := github.GetRepositories(om.orgs.target, om.orgs.targetToken)
 
 	if err != nil {
 		slog.Error("error fetching repositories from target organization")
-		return MigrationResult{}, err
+		return migrationResult{}, err
 	}
 
 	m := make(map[string]bool)
@@ -165,26 +164,26 @@ func (om OrgMigration) Migrate() (MigrationResult, error) {
 		}
 	}
 
-	migrationResult := MigrationResult{
+	mr := migrationResult{
 		Timestamp: time.Now().UTC(),
-		SourceOrg: om.source,
-		TargetOrg: om.target,
+		SourceOrg: om.orgs.source,
+		TargetOrg: om.orgs.target,
 		Migrated:  migrated,
 		Failed:    failed,
 	}
 
-	jsonData, err := json.MarshalIndent(migrationResult, "", "  ")
+	jsonData, err := json.MarshalIndent(mr, "", "  ")
 	if err != nil {
 		slog.Error("failed to parse result", err)
-		return MigrationResult{}, err
+		return migrationResult{}, err
 	}
 
-	err = github.CreateIssue(om.target, statusRepoName, "Migration result", string(jsonData), om.targetToken)
+	err = github.CreateIssue(om.orgs.target, statusRepoName, "Migration result", string(jsonData), om.orgs.targetToken)
 
 	if err != nil {
 		slog.Error("error creating issue with migration result. Check migration-result.json for details")
-		return MigrationResult{}, err
+		return migrationResult{}, err
 	}
 
-	return migrationResult, nil
+	return mr, nil
 }
