+++ internal/migration/repoMigration.go
@@ -1,12 +1,13 @@
 package migration
 
 type RepoMigration struct {
-	name                 string
-	sourceOrg, targetOrg string
+	name string
+	orgs orgs
 }
 
-func NewRepoMigration(name, sourceOrg, targetOrg string) RepoMigration {
-	return RepoMigration{name, sourceOrg, targetOrg}
+func NewRepoMigration(name, sourceOrg, targetOrg, sourceToken, targetToken string, retries int) RepoMigration {
+	maxRetries = retries
+	return RepoMigration{name, orgs{sourceOrg, targetOrg, sourceToken, targetToken}}
 }
 
 func (rm RepoMigration) Migrate() error {
